+++ ImageFeed/Profile/ProfileStructures.swift
@@ -7,20 +7,6 @@
 
 import Foundation
 
-struct ProfileResult: Codable {
-    let userLogin: String
-    let firstName: String?
-    let lastName: String?
-    let bio: String?
-    
-    private enum CodingKeys: String, CodingKey {
-        case userLogin = "username"
-        case firstName = "first_name"
-        case lastName = "last_name"
-        case bio
-    }
-}
-
 struct Profile {
     var username: String
     var name: String
